% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bind_list.R
\name{bind_list}
\alias{bind_list}
\title{rbind a List of named data.frames or vectors}
\usage{
bind_list(x, id.name = "id", content.name = "content")
}
\arguments{
\item{x}{A named \code{\link[base]{list}} of
\code{\link[base]{data.frame}}s or \code{\link[base]{vector}}}

\item{id.name}{The name to use for the column created from the \code{\link[base]{list}}.}

\item{content.name}{The name to use for the column created from the \code{\link[base]{list}}
of \code{\link[base]{vector}}s (only used if \code{x} is  \code{\link[base]{vector}}).}
}
\value{
Returns a \code{\link[data.table]{data.table}} with the \code{\link[base]{names}}
from the \code{\link[base]{list}} as an \code{id} column.
}
\description{
\code{\link[base]{rbind}} a named \code{\link[base]{list}} of
\code{\link[base]{data.frame}}s or \code{\link[base]{vector}}s to
output a single \code{\link[base]{data.frame}} with the \code{\link[base]{names}}
from the \code{\link[base]{list}} as an \code{id} column.
}
\examples{
bind_list(list(p=1:500, r=letters))
bind_list(list(p=mtcars, r=mtcars, z=mtcars, d=mtcars))
\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(rvest, magrittr)

debates <- c(
    wisconsin = "110908",
    boulder = "110906",
    california = "110756",
    ohio = "110489"
)

lapply(debates, function(x){
    txt <- read_html(paste0("http://www.presidency.ucsb.edu/ws/index.php?pid=", x)) \%>\%
        html_nodes("p") \%>\%
        html_text()

    sapply(split_index(txt, grep("^[A-Z]+:", txt)), paste, collapse = " ") \%>\%
        split_transcript() \%>\%
        split_sentence()
}) \%>\%
    bind_list("location"
}
}

