% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_matrix.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{Reorder a Matrix Based on Hierarchical Clustering}
\usage{
cluster_matrix(x, dim = "both", method = "ward.D2", ...)
}
\arguments{
\item{x}{A matrix.}

\item{dim}{The dimension to reorder (cluster); must be set to "row", "col", 
or "both".}

\item{method}{The agglomeration method to be used (see \code{\link[stats]{hclust}}).}

\item{\ldots}{ignored.}
}
\value{
Returns a reordered matrix.
}
\description{
Reorder a matrice's rows, columns, or both via hierarchical clustering.
}
\examples{
cluster_matrix(mtcars)
cluster_matrix(mtcars, dim = 'row')
cluster_matrix(mtcars, dim = 'col')
}
\seealso{
\code{\link[stats]{hclust}}
}
