% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textshape-package.R
\docType{data}
\name{golden_rules}
\alias{golden_rules}
\title{Sentence Boundary Disambiguation Edge Cases}
\format{
A data frame with 45 rows and 3 variables
}
\usage{
data(golden_rules)
}
\description{
A slightly filtered dataset containing Dias's sentence boundary
disambiguation edge cases.  This is a nested data set with the outcome
column as a nested list of desired splits.  The non-ASCII cases and spaced
ellipsis examples have been removed.
}
\details{
\itemize{
  \item Rule. The name of the rule to test
  \item Text. The testing text
  \item Outcome. The desired outcome of the sentence disambiguation
}
}
\references{
Dias, Kevin S. 2015. Golden Rules (English).
Retrieved: https://s3.amazonaws.com/tm-town-nlp-resources/golden_rules.txt
}
\keyword{datasets}
