% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textshape-package.R
\docType{data}
\name{simple_dtm}
\alias{simple_dtm}
\title{Simple \code{\link[tm]{DocumentTermMatrix}}}
\format{A list with 6 elements}
\usage{
data(simple_dtm)
}
\description{
A dataset containing a simple \code{\link[tm]{DocumentTermMatrix}}.
}
\details{
\describe{
  \item{i}{The document locations}
  \item{j}{The term locations}
  \item{v}{The count of terms for that particular element position}
  \item{nrow}{The number of rows}
  \item{ncol}{The number of columns}
  \item{dimnames}{document and terms}
}
}
\keyword{datasets}
