% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_index.R
\name{split_index}
\alias{split_index}
\alias{split_index.character}
\alias{split_index.data.frame}
\alias{split_index.default}
\alias{split_index.factor}
\alias{split_index.list}
\alias{split_index.matrix}
\alias{split_index.numeric}
\title{Split Data Forms at Specified Indices}
\usage{
split_index(x, indices, names = NULL, ...)

\method{split_index}{list}(x, indices, names = NULL, ...)

\method{split_index}{data.frame}(x, indices, names = NULL, ...)

\method{split_index}{matrix}(x, indices, names = NULL, ...)

\method{split_index}{numeric}(x, indices, names = NULL, ...)

\method{split_index}{factor}(x, indices, names = NULL, ...)

\method{split_index}{character}(x, indices, names = NULL, ...)

\method{split_index}{default}(x, indices, names = NULL, ...)
}
\arguments{
\item{x}{A data form (\code{list}, \code{vector}, \code{data.frame},
\code{matrix}).}

\item{indices}{A vector of integer indices to split at.  If \code{indices}
contains the index 1, it will be silently dropped.}

\item{names}{Optional vector of names to give to the list elements.}

\item{\ldots}{Ignored.}
}
\value{
Returns of list of data forms broken at the \code{indices}.
}
\description{
Split data forms at specified integer indices.
}
\note{
Two dimensional object will retain dimension (i.e., \code{drop = FALSE}
is used).
}
\examples{
## character
split_index(LETTERS, c(4, 10, 16))
split_index(LETTERS, c(4, 10, 16), c("dog", "cat", "chicken", "rabbit"))

## numeric
split_index(1:100, c(33, 66))

## factor
(p_chng <- head(1 + cumsum(rle(as.character(CO2[["Plant"]]))[[1]]), -1))
split_index(CO2[["Plant"]], p_chng)

## list
split_index(as.list(LETTERS), c(4, 10, 16))

## data.frame
(vs_change <- head(1 + cumsum(rle(as.character(mtcars[["vs"]]))[[1]]), -1))
split_index(mtcars, vs_change)

## matrix
(mat <- matrix(1:50, nrow=10))
split_index(mat, c(3, 6, 10))
}

