% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_tokens.R
\name{split_tokens}
\alias{split_tokens}
\alias{split_tokens.data.frame}
\alias{split_tokens.default}
\title{Split Tokens}
\usage{
split_tokens(x, ...)

\method{split_tokens}{default}(x, lower = TRUE, ...)

\method{split_tokens}{data.frame}(x, text.var = TRUE, lower = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} or character vector with tokens.}

\item{lower}{logical.  If \code{TRUE} the words are converted to lower case.}

\item{text.var}{The name of the text variable.  If missing
\code{split_tokens} tries to detect the text column with tokens.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of vectors of tokens or an expanded
\code{\link[base]{data.frame}} with tokens split apart.
}
\description{
Split tokens.
}
\examples{
(x <- c(
    "Mr. Brown comes! He says hello. i give him coffee.",
    "I'll go at 5 p. m. eastern time.  Or somewhere in between!",
    "go there"
))
split_tokens(x)
split_tokens(x, lower=FALSE)

data(DATA)
split_tokens(DATA)
split_tokens(DATA, lower=FALSE)

## Larger data set
split_tokens(hamlet)
}

