% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_words.R
\name{split_words}
\alias{split_words}
\alias{split_words.data.frame}
\alias{split_words.default}
\title{Split Words}
\usage{
split_words(x, ...)

\method{split_words}{default}(x, lower = TRUE, ...)

\method{split_words}{data.frame}(x, text.var = TRUE, lower = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} or character vector with words.}

\item{lower}{logical.  If \code{TRUE} the words are converted to lower case.}

\item{text.var}{The name of the text variable.  If missing
\code{split_words} tries to detect the text column with words.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of vectors of words or an expanded
\code{\link[data.table]{data.table}} with words split apart.
}
\description{
Split words.
}
\examples{
(x <- c(
    "Mr. Brown comes! He says hello. i give him coffee.",
    "I'll go at 5 p. m. eastern time.  Or somewhere in between!",
    "go there"
))
split_words(x)
split_words(x, lower=FALSE)

data(DATA)
split_words(DATA)
split_words(DATA, lower=FALSE)

## Larger data set
split_words(hamlet)
}

