% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_matrix.R
\name{tidy_matrix}
\alias{tidy_matrix}
\alias{tidy_adjaceny_matrix}
\title{Convert a Matrix into Tidy Form}
\usage{
tidy_matrix(x, row.name = "row", col.name = "col", value.name = "value",
  ...)

tidy_adjaceny_matrix(x, ...)
}
\arguments{
\item{x}{A matrix.}

\item{row.name}{A string to use for the row names that are now a column.}

\item{col.name}{A string to use for the column names that are now a column.}

\item{value.name}{A string to use for the values that are now a column.}

\item{\ldots}{ignored.}
}
\value{
Returns a tidied data.frame.
}
\description{
\code{tidy_matrix} - Converts matrices into a tidy data set.  Essentially, a
stacking of the matrix columns and repeating row/column names as necessary.

\code{tidy_adjaceny_matrix} - A wrapper for \code{tidy_matrix} with the 
\code{row.name}, \code{col.name}, & \code{value.name} all set to \code{"from"},
\code{"to"}, & \code{"n"}, assuming peparation for network analysis.
}
\examples{
data(simple_dtm)
tidy_matrix(as.matrix(simple_dtm), 'doc', 'term', 'n')

X <- as.matrix(simple_dtm[1:10, 1:10])
tidy_adjaceny_matrix(crossprod(X))
tidy_adjaceny_matrix(crossprod(t(X)))
}
