% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_run.R
\name{split_run}
\alias{split_run}
\alias{split_run.data.frame}
\alias{split_run.default}
\title{Split Runs}
\usage{
split_run(x, ...)

\method{split_run}{default}(x, ...)

\method{split_run}{data.frame}(x, text.var, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} or character vector with runs.}

\item{text.var}{The name of the text variable.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of vectors of runs or an expanded
\code{\link[data.table]{data.table}} with runs split apart.
}
\description{
Split runs of consecutive characters.
}
\examples{
x1 <- c(
     "122333444455555666666",
     NA,
     "abbcccddddeeeeeffffff",
     "sddfg",
     "11112222333"
)

x <- c(rep(x1, 2), ">>???,,,,....::::;[[")

split_run(x)


DATA[["run.col"]] <- x
split_run(DATA, "run.col")
}

