% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sentence.R
\name{split_sentence}
\alias{split_sentence}
\alias{split_sentence.default}
\alias{split_sentence.data.frame}
\title{Split Sentences}
\usage{
split_sentence(x, ...)

\method{split_sentence}{default}(x, ...)

\method{split_sentence}{data.frame}(x, text.var = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} or character vector with sentences.}

\item{text.var}{The name of the text variable.  If \code{TRUE}
\code{split_sentence} tries to detect the column with sentences.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of vectors of sentences or a expanded
\code{\link[base]{data.frame}} with sentences split apart.
}
\description{
Split sentences.
}
\examples{
(x <- paste0(
    "Mr. Brown comes! He says hello. i give him coffee.  i will ",
    "go at 5 p. m. eastern time.  Or somewhere in between!go there"
))
split_sentence(x)

data(DATA)
split_sentence(DATA)

y <- c(paste(
    "\\x91He has asked the Administration to be sent there,\\x92 said the",
    "other, \\x91with the idea of showing what he could do; and I was instructed",
    "accordingly.\\x92 They both agreed it was frightful, then made several",
    "bizarre remarks: \\x91Make rain and fine weather\\x-one man\\x-the Council\\x-by the",
    "nose\\x92\\x-bits of absurd sentences that got the better of my drowsiness when",
    "the uncle said, \\x91The climate may do away with this difficulty for you.",
    "And one more, \\x93How bout that!\\x94 But still there is \\x93another.\\x94,",
    "but who?  No. 3 will.  No.  He will not!",
    collapse = ' '
), "I said no.  Now stop!", "I will not!  Yes you will.")
Encoding(y) <- "latin1"
y
## doesn't work as expected because of encoding issue
split_sentence(y)

\dontrun{
library(textclean)
replace_curly <- function(x, ...){
    replaces <- c('\\x91', '\\x92', '\\x93', '\\x94')
    Encoding(replaces) <- "latin1"
    textclean::mgsub(x, replaces, c("'", "'", "\\"", "\\""))
}
split_sentence(replace_curly(y))
}
}
