% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_transcript.R
\name{split_transcript}
\alias{split_transcript}
\title{Split a Transcript Style Vector on Deliminator & Coerce to Dataframe}
\usage{
split_transcript(x, delim = ":", colnames = c("person", "dialogue"))
}
\arguments{
\item{x}{A transcript style vector (e.g., \code{c("greg: Who me", "sarah: yes you!")}.}

\item{delim}{The deliminator to split on.}

\item{colnames}{The column names to use for the \code{\link[data.table]{data.table}} output.}
}
\value{
Returns a \code{\link[data.table]{data.table}}.
}
\description{
Split a transcript style vector (e.g., \code{c("greg: Who me", "sarah: yes you!")}
into a name and dialogue vector that is coerced to a \code{\link[data.table]{data.table}}.
Leading/trailing white space in the columns is stripped out.
}
\examples{
split_transcript(c("greg: Who me", "sarah: yes you!"))
}

