% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_transcript.R
\name{split_transcript}
\alias{split_transcript}
\title{Split a Transcript Style Vector on Delimitor & Coerce to Dataframe}
\usage{
split_transcript(x, delim = ":", colnames = c("person", "dialogue"))
}
\arguments{
\item{x}{A transcript style vector (e.g., \code{c("greg: Who me", "sarah: yes you!")}.}

\item{delim}{The delimitor to split on.}

\item{colnames}{The column names to use for the \code{\link[data.table]{data.table}} output.}
}
\value{
Returns a 2 column \code{\link[data.table]{data.table}}.
}
\description{
Split a transcript style vector (e.g., \code{c("greg: Who me", "sarah: yes you!")}
into a name and dialogue vector that is coerced to a \code{\link[data.table]{data.table}}.
Leading/trailing white space in the columns is stripped out.
}
\examples{
split_transcript(c("greg: Who me", "sarah: yes you!"))

\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(rvest, magrittr, xml2)

debates <- c(
    wisconsin = "110908",
    boulder = "110906",
    california = "110756",
    ohio = "110489"
)

lapply(debates, function(x){
    txt <- read_html(paste0("http://www.presidency.ucsb.edu/ws/index.php?pid=", x)) \%>\%
        html_nodes("p") \%>\%
        html_text()

    sapply(split_index(txt, grep("^[A-Z]+:", txt)), paste, collapse = " ") \%>\%
        split_transcript() \%>\%
        split_sentence()
})
}
}

