% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textshape-package.R
\docType{package}
\name{textshape}
\alias{package-textshape}
\alias{textshape}
\alias{textshape-package}
\title{What the Package Does (one line)}
\description{
What the package does (one or more paragraphs)
}

